fate_flow: rename fate_flow/operator to fate_flow/operators
